requirements.txt
# OpenAI Agents Python SDK with voice support
openai-agents[voice]>=0.1.0

# Audio processing dependencies
numpy>=1.24.0
sounddevice>=0.4.6

# Additional dependencies for audio handling
scipy>=1.10.0
librosa>=0.10.0

# Web framework for API integration
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
